+++ CHANGELOG.md
@@ -38,6 +38,7 @@ Changelog](https://keepachangelog.com/en/1.0.0/).
 - SNMP exploiter plugin. #4258
 - Zerologon exploiter plugin. #4259
 - Mimikatz credentials collector plugin. #4260
+- Chrome credentials collector plugin. #4261
 
 ### Security